+++ index.commonjs.js
@@ -2,4 +2,5 @@
 
 exports.Form = require('./commonjs/form').default
 exports.Field = require('./commonjs/field').default
+exports.List = require('./commonjs/list').default
 exports.Submit = require('./commonjs/submit').default