+++ index.js
@@ -1,3 +1,4 @@
 export { default as Form } from './modules/form'
 export { default as Field } from './modules/field'
+export { default as List } from './modules/list'
 export { default as Submit } from './modules/submit'